refactor(RandomRacingStrategy): Move directory
racing 하위 폴더에 위치했습니다.
